task6

+++ main.h
@@ -39,7 +39,7 @@ int print_oct(va_list args);
 void putoct(unsigned int num);
 int octlen(unsigned int num);
 int print_hex(va_list args);
-void puthex(unsigned int num);
+void puthex(unsigned long int num);
 int hexlen(unsigned int num);
 int print_HEX(va_list args);
 void putHEX(unsigned int num);
